requirements.txt
Django==1.9
django-extensions==1.6.1
djangorestframework==3.3.2
drf-nested-routers==0.11.1
parse==1.6.6
PyYAML==3.11
six==1.10.0
